backend/requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.23
sqlalchemy>=2.0
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
email-validator>=2.0
